feat: persentase resiko

+++ app/src/main/java/com/itb/diabetify/presentation/whatif/WhatIfResultScreen.kt
@@ -89,7 +89,7 @@ fun WhatIfResultScreen(
             HomeCard(
                 title = "Persentase Risiko",
                 hasWarning = true,
-                riskPercentage = predictionPercentage.toFloat()
+                riskPercentage = predictionPercentage
             ) {
                 Column(
                     modifier = Modifier
